requirements.txt
Flask
pickle-mixin
faiss-cpu
langchain_groq
langchain
langchain-community
langchain-google-genai
transformers
torch
